requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==41.0.4
dataclasses-json==0.5.14
distro==1.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.104.0
filelock==3.12.4
Flask==2.3.3
Flask-Sitemap==0.4.0
frozenlist==1.4.0
fsspec==2023.9.2
gevent==23.9.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
html-sanitizer==2.2.0
httplib2==0.20.2
huggingface-hub==0.17.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
install==1.3.5
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langchain==0.0.279
langsmith==0.0.40
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
oauthlib==3.2.0
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
passlib==1.7.4
pdfminer.six==20221105
pexpect==4.8.0
Pillow==9.5.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyrsistent==0.18.1
pyserial==3.5
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.24
python-multipart==0.0.6
python-pptx==0.6.22
pytz==2022.1
redis==5.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
SecretStorage==3.3.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
ssh-import-id==5.11
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
Twisted==22.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.5
uvicorn==0.23.2
wadllib==1.3.6
Werkzeug==2.3.7
XlsxWriter==3.1.7
yarl==1.9.2
zipp==1.0.0
zope.event==5.0
zope.interface==5.4.0
